requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0

# LLM Integration
openai==1.3.0
google-generativeai==0.3.0
requests==2.31.0

# Audio/Transcription
pydub==0.25.1

# PDF Export (optional)
reportlab==4.0.7
Pillow==10.1.0

# Database
django-filter==23.4

# Development
django-extensions==3.2.3
Werkzeug==3.0.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
